requirements.txt
nmap
schedule
requests
vulners
python-dotenv
